minimal misc revisions
