+++ docker-compose.yaml
@@ -12,4 +12,4 @@ services:
         max-size: "10M"
         max-file: "10"
     env_file:
-      - object-ledger-wo-mqtt.env
+      - object-ledger.env
