+++ object-ledger.env
@@ -1,10 +1,13 @@
 MQTT_IP=mqtt
 HOSTNAME=skyscan-011
-LATITUDE_L=0.0
-LONGITUDE_L=0.0
-ALTITUDE_L=0.0
 CONFIG_TOPIC=/Multimodal/${HOSTNAME}/config/edgetech-object-ledger/JSON
 ADS_B_INPUT_TOPIC=/Multimodal/${HOSTNAME}/ADS-B/edgetech-dump1090-json/JSON
 AIS_INPUT_TOPIC=/Multimodal/${HOSTNAME}/AIS/edgetech-daisy/JSON
-LEGDER_OUTPUT_TOPIC=/Multimodal/${HOSTNAME}/ledger/edgetech-axis-ptz-controller/JSON
-SELECTION_OUTPUT_TOPIC=/Multimodal/${HOSTNAME}/selection/edgetech-axis-ptz-controller/JSON
+LEDGER_OUTPUT_TOPIC=/Multimodal/${HOSTNAME}/ledger/edgetech-axis-ptz-controller/JSON
+MAX_AIRCRAFT_ENTRY_AGE=60.0
+MAX_SHIP_ENTRY_AGE=180.0
+PUBLISH_INTERVAL=1
+HEARTBEAT_INTERVAL=10
+LOOP_SLEEP=0.1
+CONTINUE_ON_EXCEPTION=False
+LOG_LEVEL="DEBUG"
