+++ object-ledger/pyproject.toml
@@ -8,8 +8,6 @@ license = "Apache 2.0"
 [tool.poetry.dependencies]
 python = "^3.11"
 coloredlogs = "^15.0.1"
-numpy = "^1.24.2"
-numpy-quaternion = "^2022.4.3"
 paho-mqtt = "^1.5.0"
 pandas = "^2.0.2"
 schedule = "^1.2.0"
